Inline basic baking code to bake.html; Extract core to bakeUtils
